Move Location and Ride to BL namespace. Refactor code to reflect change.
