+++ TaxiDispatcher.App/Scheduler.cs
@@ -1,6 +1,6 @@
-﻿using TaxiDispatcher.App;
-using TaxiDispatcher.BL.CustomExceptions;
+﻿using TaxiDispatcher.BL.CustomExceptions;
 using TaxiDispatcher.BL.Extensions;
+using TaxiDispatcher.BL.Locations;
 using TaxiDispatcher.BL.Rides;
 using TaxiDispatcher.BL.Taxis;
 using TaxiDispatcher.DAL;
