+++ TaxiDispatcher.App/TaxiDispatcher.BL.csproj
@@ -44,10 +44,10 @@
     <Compile Include="CustomExceptions\InvalidTaxiCompanyException.cs" />
     <Compile Include="CustomExceptions\NoAvailableTaxiVehiclesException.cs" />
     <Compile Include="Extensions\EnumerableExtensions.cs" />
-    <Compile Include="Location.cs" />
+    <Compile Include="Locations\Location.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Rides\RideFactory.cs" />
-    <Compile Include="RideOrder.cs" />
+    <Compile Include="Rides\RideOrder.cs" />
     <Compile Include="Rides\DayCityRide.cs" />
     <Compile Include="Rides\DayInterCityRide.cs" />
     <Compile Include="Rides\NightCityRide.cs" />
