+++ TaxiDispatcher.Client.Tests/AppTests/SelectingClosestTaxi/RideOrderSelectsClosestTaxiTests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using TaxiDispatcher.App;
 using TaxiDispatcher.BL;
 using TaxiDispatcher.BL.CustomExceptions;
+using TaxiDispatcher.BL.Rides;
 using TaxiDispatcher.Tests.HelperClasses;
 using Xunit;
